Rename module-common-react to module-common
