+++ modules/module/client-react/package.json
@@ -1,6 +1,8 @@
 {
   "name": "@module/module-client-react",
   "version": "1.0.0",
-  "private": true
+  "private": true,
+  "dependencies": {
+    "@module/module-common": "^1.0.0"
+  }
 }
-
