+++ modules/module/common/package.json
@@ -1,9 +1,8 @@
 {
-  "name": "@module/module-common-react",
+  "name": "@module/module-common",
   "version": "1.0.0",
   "private": true,
   "dependencies": {
     "fractal-objects": "^0.10.0"
   }
 }
-
