+++ modules/module/server-ts/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@module/module-common-react": "^1.0.0"
+    "@module/module-common": "^1.0.0"
   }
 }
