+++ modules/validation/common-react/__tests__/Validation.spec.ts
@@ -2,7 +2,7 @@
 import { expect } from 'chai';
 import { step } from 'mocha-steps';
 import i18next from 'i18next';
-import CommonModule from '@module/module-common-react';
+import CommonModule from '@module/module-common';
 import i18n from '@module/i18n-common-react';
 import validation from '..';
