+++ modules/validation/common-react/package.json
@@ -1,5 +1,8 @@
 {
   "name": "@module/validation-common-react",
   "version": "1.0.0",
-  "private": true
+  "private": true,
+  "dependencies": {
+    "@module/module-common": "^1.0.0"
+  }
 }
